Power control
  * change to floating point control to allow for
    fractional changes in power level, i.e. 0.1 W / step

+++ src/include/elecraft/K2.h
@@ -68,7 +68,7 @@ public:
 	int  get_PTT();
 
 	void get_pc_min_max_step(double &min, double &max, double &step);
-	int  get_power_control();
+	double get_power_control();
 	void set_power_control(double val);
 	int  get_smeter();
 	int  get_power_out();
