+++ src/include/kenwood/TS990.h
@@ -136,7 +136,7 @@ public:
 	int  get_volume_control();
 
 	void set_power_control(double val);
-	int  get_power_control(void);
+	double get_power_control(void);
 	void get_pc_min_max_step(double &min, double &max, double &step) {
 		min = 0; pmax = max = 200; step = 1; }
 	int  get_power_out();
