+++ src/include/other/TMD710.h
@@ -33,7 +33,7 @@ public:
 //	int  get_smeter();
 //	int  get_swr();
 //	int  get_power_out();
-//	int  get_power_control();
+//	double get_power_control();
 //	void set_volume_control(double val);
 //	void set_power_control(double val);
 	void set_PTT_control(int val);
@@ -41,7 +41,7 @@ public:
 //	void tune_rig();
 //	void set_bwA(int val);
 
-	int get_power_control();
+	double get_power_control();
 	void set_power_control(double val);
 	void get_pc_min_max_step(double &min, double &max, double &step) {
 		min = 5; max = 50; step = 5; }
