+++ src/include/tentec/TT550.h
@@ -114,7 +114,7 @@ public:
 
 	int  get_power_out();
 
-	int  get_power_control();
+	double get_power_control();
 	void set_power_control(double val);
 	void get_pc_min_max_step(double &min, double &max, double &step) {
 		min = 5; pmax = max = 100; step = 1; }
