+++ src/rigs/icom/IC705.cxx
@@ -1286,7 +1286,7 @@ void RIG_IC705::set_power_control(double val)
 	waitFB("set power");
 }
 
-int RIG_IC705::get_power_control()
+double RIG_IC705::get_power_control()
 {
 	int val = progStatus.power_level;
 	string cstr = "\x14\x0A";
