+++ src/rigs/icom/IC7200.cxx
@@ -901,7 +901,7 @@ void RIG_IC7200::set_power_control(double val)
 	seth();
 }
 
-int RIG_IC7200::get_power_control()
+double RIG_IC7200::get_power_control()
 {
 	string cstr = "\x14\x0A";
 	string resp = pre_fm;
