+++ src/rigs/icom/IC756.cxx
@@ -759,7 +759,7 @@ void RIG_IC756PRO::set_power_control(double val)
 	waitFB("set power");
 }
 
-int RIG_IC756PRO::get_power_control()
+double RIG_IC756PRO::get_power_control()
 {
 	string cstr = "\x14\x0A";
 	string resp = pre_fm;
