+++ src/rigs/icom/IC756PRO3.cxx
@@ -674,7 +674,7 @@ void RIG_IC756PRO3::set_power_control(double val)
 	waitFB("set power");
 }
 
-int RIG_IC756PRO3::get_power_control()
+double RIG_IC756PRO3::get_power_control()
 {
 	string cstr = "\x14\x0A";
 	string resp = pre_fm;
