+++ src/rigs/icom/IC9700.cxx
@@ -994,7 +994,7 @@ void RIG_IC9700::set_power_control(double val)
 	set_trace(2, "set_power_control()", str2hex(replystr.c_str(), replystr.length()));
 }
 
-int RIG_IC9700::get_power_control()
+double RIG_IC9700::get_power_control()
 {
 	int val = progStatus.power_level;
 	string cstr = "\x14\x0A";
