+++ src/rigs/kenwood/TS590SG.cxx
@@ -331,7 +331,7 @@ void RIG_TS590SG::set_power_control(double val)
 	sendCommand(cmd, 0);
 }
 
-int RIG_TS590SG::get_power_control()
+double RIG_TS590SG::get_power_control()
 {
 	cmd = "PC;";
 	get_trace(1, "get_power_control");
