+++ src/rigs/yaesu/FT5000.cxx
@@ -503,7 +503,7 @@ int RIG_FT5000::get_power_out()
 }
 
 // Transceiver power level
-int RIG_FT5000::get_power_control()
+double RIG_FT5000::get_power_control()
 {
 	cmd = rsp = "PC";
 	cmd += ';';
