+++ src/rigs/yaesu/FT950.cxx
@@ -521,7 +521,7 @@ int RIG_FT950::get_power_out()
 }
 
 // Transceiver power level
-int RIG_FT950::get_power_control()
+double RIG_FT950::get_power_control()
 {
 	cmd = rsp = "PC";
 	cmd += ';';
