+++ src/rigs/yaesu/FT991.cxx
@@ -469,7 +469,7 @@ int RIG_FT991::get_power_out()
 }
 
 // Transceiver power level
-int RIG_FT991::get_power_control()
+double RIG_FT991::get_power_control()
 {
 	cmd = rsp = "PC";
 	cmd += ';';
