+++ src/rigs/yaesu/FT991A.cxx
@@ -487,7 +487,7 @@ int RIG_FT991A::get_power_out()
 }
 
 // Transceiver power level
-int RIG_FT991A::get_power_control()
+double RIG_FT991A::get_power_control()
 {
 	cmd = rsp = "PC";
 	cmd += ';';
