+++ src/rigs/yaesu/FTdx1200.cxx
@@ -517,7 +517,7 @@ int RIG_FTdx1200::get_power_out()
 }
 
 // Transceiver power level
-int RIG_FTdx1200::get_power_control()
+double RIG_FTdx1200::get_power_control()
 {
 	cmd = rsp = "PC";
 	cmd += ';';
